Generate seeds

+++ db/migrate/20211115162011_create_foods.rb
@@ -3,7 +3,7 @@ class CreateFoods < ActiveRecord::Migration[6.1]
     create_table :foods do |t|
       t.string :name
       t.string :measure_unit
-      t.integer :price
+      t.decimal :price
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
